split sidebar in own components
